@@ -7,7 +7,7 @@
   "scripts": {
     "dev": "nuxt",
     "build": "nuxt build",
-    "start": "nuxt start",
+    "start": "npm run build && nuxt start",
     "generate": "nuxt generate",
     "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
     "test": "jest"
@@ -19,13 +19,17 @@
     "@fortawesome/vue-fontawesome": "^0.1.8",
     "@nuxtjs/axios": "^5.3.6",
     "@types/lodash": "^4.14.149",
+    "bcryptjs": "^2.4.3",
     "bootstrap": "^4.1.3",
     "bootstrap-vue": "^2.0.0",
+    "identity-obj-proxy": "^3.0.0",
     "nuxt": "^2.10.2",
     "nuxt-vuex-localstorage": "^1.2.6",
     "vee-validate": "^3.2.1",
     "vue-property-decorator": "^8.3.0",
     "vue-sidebar-menu": "^4.4.3",
+    "vue-slider-component": "^3.0.44",
+    "vue-star-rating": "^1.6.1",
     "vue-template-loader": "^1.0.0",
     "webpack-stream": "^5.2.1"
   },
@@ -37,7 +41,7 @@
     "@vue/test-utils": "^1.0.0-beta.27",
     "babel-eslint": "^10.0.1",
     "babel-jest": "^24.1.0",
-    "eslint": "^6.1.0",
+    "eslint": "^6.7.2",
     "eslint-plugin-nuxt": ">=0.4.2",
     "jest": "^24.1.0",
     "vue-jest": "^4.0.0-0"
