@@ -1,7 +1,9 @@
 import express from 'express';
 const router = express.Router();
 import Message from '../../types/message';
-const $messages_methods = require('./messages_methods')
+import { MessageController } from './messages_methods';
+
+let messagecontroller: MessageController = new MessageController();
 
 /**
  * Middleware that is specific to this router
@@ -22,7 +24,7 @@ router.get('/conversation/:user_id1/:user_id2', async (req: any, res: any) => {
     const userID1: number = parseInt(req.params.user_id1);
     const userID2: number = parseInt(req.params.user_id2);
     try{
-        const messages: Message[] = await $messages_methods.getConversation(userID1, userID2);
+        const messages: Message[] = await messagecontroller.getConversation(userID1, userID2);
         res.status(200).json(messages);
     } catch (err) {
         const statusCode: number = parseInt(err);
@@ -37,7 +39,7 @@ router.get('/conversation/:user_id1/:user_id2', async (req: any, res: any) => {
 router.get('/:user_id', async (req: any, res: any) => {
     const userID: number = parseInt(req.params.user_id);
     try{
-        const userIDs: number[] = await $messages_methods.getCorrespondents(userID);
+        const userIDs: number[] = await messagecontroller.getCorrespondents(userID);
         res.status(200).json(userIDs);
     } catch (err) {
         const statusCode: number = parseInt(err);
@@ -53,7 +55,7 @@ router.get('/:user_id', async (req: any, res: any) => {
 router.post('/', async (req: any, res: any) => {
     const msg: Message = req.body.message;
     try{
-        const newMessageID: number = await $messages_methods.addMessage(msg);
+        const newMessageID: number = await messagecontroller.addMessage(msg);
         res.status(200).json({id: newMessageID});
     } catch (err) {
         const statusCode: number = parseInt(err);
