@@ -2,7 +2,9 @@ import express from 'express';
 const router = express.Router();
 import Project from '../../types/project';
 import Category from '../../types/category';
-const $projects_categories_methods = require("./projects_categories_methods");
+import { ProjectCategoryController } from './projects_categories_methods';
+
+let projectcategorycontroller: ProjectCategoryController = new ProjectCategoryController();
 
 /**
  * Middleware that is specific to this router
@@ -22,7 +24,7 @@ router.use((req: any, res: any, next: Function) => {
 router.get('/project/:project_id', async (req: any, res: any) => {
     const projectID: number = parseInt(req.params.project_id);
     try{
-        const categories: Category[] = await $projects_categories_methods.getAllCategoriesOfProject(projectID);
+        const categories: Category[] = await projectcategorycontroller.getAllCategoriesOfProject(projectID);
         res.status(200).json(categories);
     } catch (err) {
         const statusCode: number = parseInt(err);
@@ -37,7 +39,7 @@ router.get('/project/:project_id', async (req: any, res: any) => {
 router.get('/category/:category_id', async (req: any, res: any) => {
     const categoryID: number = parseInt(req.params.category_id);
     try{
-        const projects: Project[] = await $projects_categories_methods.getAllProjectsOfCategory(categoryID);
+        const projects: Project[] = await projectcategorycontroller.getAllProjectsOfCategory(categoryID);
         res.status(200).json(projects);
     } catch (err) {
         const statusCode: number = parseInt(err);
@@ -53,7 +55,7 @@ router.post('/:project_id/:category_id', async (req: any, res: any) => {
     const projectID: number = parseInt(req.params.project_id);
     const categoryID: number = parseInt(req.params.category_id);
     try{
-        await $projects_categories_methods.addProjectToCategory(projectID, categoryID);
+        await projectcategorycontroller.addProjectToCategory(projectID, categoryID);
         res.status(200).send("Successfully assigned project to category.");
     } catch (err) {
         const statusCode: number = parseInt(err);
@@ -70,7 +72,7 @@ router.delete('/:project_id/:category_id', async (req: any, res: any) => {
     const projectID: number = parseInt(req.params.project_id);
     const categoryID: number = parseInt(req.params.category_id);
     try{
-        await $projects_categories_methods.removeProjectFromCategory(projectID, categoryID);
+        await projectcategorycontroller.removeProjectFromCategory(projectID, categoryID);
         res.status(200).send("Successfully removed project from category.");
     } catch (err) {
         const statusCode: number = parseInt(err);
