@@ -1,10 +1,11 @@
 import Skill from "./skill";
 
 interface Profile {
-    id: number|null;
+    id: number;
     name: string;
     project_id: number;
     skills: Skill[];
+    questions: string[];
 }
 
 export default Profile;
