@@ -1,6 +1,7 @@
 interface Category {
-    id: number|null;
+    id: number;
     name: string;
+    subcategory: string|null;
 }
 
 export default Category;
