@@ -1,8 +1,10 @@
-interface Notification {
+interface INotification {
     id: number|null;
     user_id: number;
     status: number;
     dest_url: string;
+    msg: string;
+    created_at: string;
 }
 
-export default Notification;
+export default INotification;
