@@ -1,9 +1,10 @@
 import Profile from './profile';
+import Category from './category'
 
 interface Project {
     id: number|null;
     name: string;
-    status: number;
+    status: number; // 0: just created looking for members, 1: looking for members, 2: found all members, awaiting approval, 3: found all members, approved
     pitch: string;
     created_at: string; // 'YYYY-MM-DD hh:mm:ss' format
     edited_at: string; // 'YYYY-MM-DD hh:mm:ss' format
@@ -11,6 +12,7 @@ interface Project {
     creator_first_name: string;
     creator_last_name: string;
     profiles: Profile[];
+    categories: Category[];
 }
 
 export default Project;
